refactor: standardise game number types to u32, eliminating as-casts
Board numbers, target, builder input, and error variants all use u32
now. Count parameters (small_number_count, large_number_count) remain
u8 as they represent counts 0-6, not game numbers. This removes all
narrowing/widening casts (as u8, as u32, as u16, as usize) from the
game and solver logic.

+++ benches/solver_benchmark.rs
@@ -5,10 +5,10 @@ use yew_countdown_solver::solver::iterative_deepening::IterativeDeepeningSolver;
 use yew_countdown_solver::solver::traits::Solver;
 
 /// Helper to build a Game from a target and a slice of numbers.
-fn make_game(target: u16, numbers: &[u32]) -> Game {
+fn make_game(target: u32, numbers: &[u32]) -> Game {
     let mut builder = BoardBuilder::new();
     for &n in numbers {
-        builder = builder.add_number(n as u8).unwrap();
+        builder = builder.add_number(n).unwrap();
     }
     let board = builder.build().unwrap();
     Game::new(board, target).unwrap()
